Time: 4 ms (61.13%) | Memory: 9.2 MB (13.59%) - LeetSync

+++ 3055-maximum-odd-binary-number/maximum-odd-binary-number.cpp
@@ -3,7 +3,7 @@ public:
     string maximumOddBinaryNumber(string s) {
 
         string ones, zeros;
-        int len = s.length();
+        short int len = s.length();
         bool oneFound = false;
 
         for (int i = 0; i < len; i++)
